+++ source/MaterialXCore/Look.cpp
@@ -30,7 +30,51 @@ MaterialAssignPtr Look::addMaterialAssign(const string& name, const string& mate
     return matAssign;
 }
 
-void Look::setInheritsFrom(LookPtr look)
+vector<MaterialAssignPtr> Look::getActiveMaterialAssigns() const
+{
+    vector<MaterialAssignPtr> activeAssigns;
+    for (ConstElementPtr elem : traverseInheritance())
+    {
+        vector<MaterialAssignPtr> assigns = elem->asA<Look>()->getMaterialAssigns();
+        activeAssigns.insert(activeAssigns.end(), assigns.begin(), assigns.end());
+    }
+    return activeAssigns;
+}
+
+vector<PropertyAssignPtr> Look::getActivePropertyAssigns() const
+{
+    vector<PropertyAssignPtr> activeAssigns;
+    for (ConstElementPtr elem : traverseInheritance())
+    {
+        vector<PropertyAssignPtr> assigns = elem->asA<Look>()->getPropertyAssigns();
+        activeAssigns.insert(activeAssigns.end(), assigns.begin(), assigns.end());
+    }
+    return activeAssigns;
+}
+
+vector<PropertySetAssignPtr> Look::getActivePropertySetAssigns() const
+{
+    vector<PropertySetAssignPtr> activeAssigns;
+    for (ConstElementPtr elem : traverseInheritance())
+    {
+        vector<PropertySetAssignPtr> assigns = elem->asA<Look>()->getPropertySetAssigns();
+        activeAssigns.insert(activeAssigns.end(), assigns.begin(), assigns.end());
+    }
+    return activeAssigns;
+}
+
+vector<VisibilityPtr> Look::getActiveVisibilities() const
+{
+    vector<VisibilityPtr> activeVisibilities;
+    for (ConstElementPtr elem : traverseInheritance())
+    {
+        vector<VisibilityPtr> visibilities = elem->asA<Look>()->getVisibilities();
+        activeVisibilities.insert(activeVisibilities.end(), visibilities.begin(), visibilities.end());
+    }
+    return activeVisibilities;
+}
+
+void Look::setInheritsFrom(ElementPtr look)
 {
     for (LookInheritPtr inherit : getLookInherits())
     {
@@ -42,12 +86,12 @@ void Look::setInheritsFrom(LookPtr look)
     }
 }
 
-LookPtr Look::getInheritsFrom() const
+ElementPtr Look::getInheritsFrom() const
 {
     vector<LookInheritPtr> inherits = getLookInherits();
     if (inherits.empty())
     {
-        return LookPtr();
+        return nullptr;
     }
     return getRoot()->getChildOfType<Look>(inherits[0]->getName());
 }